chore(python): clarify how multi-file imports work
