+++ sdk/python/dev/dagger.json
@@ -17,5 +17,5 @@
     }
   ],
   "source": ".",
-  "engineVersion": "v0.12.7"
+  "engineVersion": "v0.13.0"
 }
