+++ sdk/python/dev/pyproject.toml
@@ -6,7 +6,7 @@ dependencies = ["dagger-io"]
 
 [tool.dagger]
 # Ensure the dev module works both with released and dev engine.
-uv-version = "0.4.9"
+uv-version = "0.4.10"
 
 [tool.uv.sources]
 dagger-io = { path = "sdk", editable = true }
