+++ sdk/python/runtime/template/main.py
@@ -13,14 +13,16 @@ rest is a long description with more detail on the module's purpose or usage,
 if appropriate. All modules should have a short description.
 """
 
-import dagger
-from dagger import dag, function, object_type
-
 # NOTE: it's recommended to move your code into other files in this package
 # and keep __init__.py for imports only, according to Python's convention.
 # The only requirement is that Dagger needs to be able to import a package
-# called "main", so as long as the files are imported here, they should be
-# available to Dagger.
+# called "main" (i.e., src/main/).
+#
+# For example, to import from src/main/main.py:
+# >>> from .main import Main as Main
+
+import dagger
+from dagger import dag, function, object_type
 
 
 @object_type
